% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{gmm_algo}
\alias{gmm_algo}
\title{Gaussian Mixture Model Algorithm}
\usage{
gmm_algo(data, K, max_iterations)
}
\arguments{
\item{data}{input data that needs to be clustered [vector of values of type double]}

\item{K}{number of clusters [integer value]}

\item{max_iterations}{number of iterations [integer value]}
}
\value{
vector of mean, covariance and prior weight values
}
\description{
This function takes in the data and returns the final estimates of mean, covariance 
and prior weight values that maximize the likelihood using Gaussian Mixture Model Algorithm
}
\examples{
 N <- 100
 K <- 2 
 max_iterations <- 100
 data <- numeric(N)
 set.seed(123)
 data <- runif(N)
 gmm_algo(data, K, max_iterations)
}
